signalfxexporter: refactor conversion code for metrics to signalfx
This PR only moves the code to a different file, and fixes few nits.
