+++ exporter/signalfxexporter/internal/translation/converter.go
@@ -16,7 +16,6 @@ package translation // import "github.com/open-telemetry/opentelemetry-collector
 
 import (
 	"fmt"
-	"math"
 	"strconv"
 	"strings"
 	"time"
@@ -38,13 +37,6 @@ var (
 	sfxMetricTypeGauge             = sfxpb.MetricType_GAUGE
 	sfxMetricTypeCumulativeCounter = sfxpb.MetricType_CUMULATIVE_COUNTER
 	sfxMetricTypeCounter           = sfxpb.MetricType_COUNTER
-
-	// Some standard dimension keys.
-	// upper bound dimension key for histogram buckets.
-	upperBoundDimensionKey = "upper_bound"
-
-	// infinity bound dimension value is used on all histograms.
-	infinityBoundSFxDimValue = float64ToDimValue(math.Inf(1))
 )
 
 // MetricsConverter converts MetricsData to sfxpb DataPoints. It holds an optional
@@ -88,7 +80,8 @@ func (c *MetricsConverter) MetricDataToSignalFxV2(rm pdata.ResourceMetrics) []*s
 	for j := 0; j < rm.InstrumentationLibraryMetrics().Len(); j++ {
 		ilm := rm.InstrumentationLibraryMetrics().At(j)
 		for k := 0; k < ilm.Metrics().Len(); k++ {
-			dps := c.metricToSfxDataPoints(ilm.Metrics().At(k), extraDimensions)
+			dps := fromMetric(ilm.Metrics().At(k), extraDimensions)
+			dps = c.translateAndFilter(dps)
 			sfxDatapoints = append(sfxDatapoints, dps...)
 		}
 	}
@@ -96,26 +89,7 @@ func (c *MetricsConverter) MetricDataToSignalFxV2(rm pdata.ResourceMetrics) []*s
 	return c.datapointValidator.sanitizeDataPoints(sfxDatapoints)
 }
 
-func (c *MetricsConverter) metricToSfxDataPoints(metric pdata.Metric, extraDimensions []*sfxpb.Dimension) []*sfxpb.DataPoint {
-	// TODO: Figure out some efficient way to know how many datapoints there
-	// will be in the given metric.
-	var dps []*sfxpb.DataPoint
-
-	basePoint := makeBaseDataPoint(metric)
-
-	switch metric.DataType() {
-	case pdata.MetricDataTypeNone:
-		return nil
-	case pdata.MetricDataTypeGauge:
-		dps = convertNumberDatapoints(metric.Gauge().DataPoints(), basePoint, extraDimensions)
-	case pdata.MetricDataTypeSum:
-		dps = convertNumberDatapoints(metric.Sum().DataPoints(), basePoint, extraDimensions)
-	case pdata.MetricDataTypeHistogram:
-		dps = convertHistogram(metric.Histogram().DataPoints(), basePoint, extraDimensions)
-	case pdata.MetricDataTypeSummary:
-		dps = convertSummaryDataPoints(metric.Summary().DataPoints(), metric.Name(), extraDimensions)
-	}
-
+func (c *MetricsConverter) translateAndFilter(dps []*sfxpb.DataPoint) []*sfxpb.DataPoint {
 	if c.metricTranslator != nil {
 		dps = c.metricTranslator.TranslateDataPoints(c.logger, dps)
 	}
@@ -135,189 +109,6 @@ func (c *MetricsConverter) metricToSfxDataPoints(metric pdata.Metric, extraDimen
 	return dps
 }
 
-func attributesToDimensions(attributes pdata.AttributeMap, extraDims []*sfxpb.Dimension) []*sfxpb.Dimension {
-	dimensions := make([]*sfxpb.Dimension, len(extraDims), attributes.Len()+len(extraDims))
-	copy(dimensions, extraDims)
-	if attributes.Len() == 0 {
-		return dimensions
-	}
-	dimensionsValue := make([]sfxpb.Dimension, attributes.Len())
-	pos := 0
-	attributes.Range(func(k string, v pdata.AttributeValue) bool {
-		dimensionsValue[pos].Key = k
-		dimensionsValue[pos].Value = v.AsString()
-		dimensions = append(dimensions, &dimensionsValue[pos])
-		pos++
-		return true
-	})
-	return dimensions
-}
-
-func convertSummaryDataPoints(
-	in pdata.SummaryDataPointSlice,
-	name string,
-	extraDims []*sfxpb.Dimension,
-) []*sfxpb.DataPoint {
-	out := make([]*sfxpb.DataPoint, 0, in.Len())
-
-	for i := 0; i < in.Len(); i++ {
-		inDp := in.At(i)
-
-		dims := attributesToDimensions(inDp.Attributes(), extraDims)
-		ts := timestampToSignalFx(inDp.Timestamp())
-
-		countPt := sfxpb.DataPoint{
-			Metric:     name + "_count",
-			Timestamp:  ts,
-			Dimensions: dims,
-			MetricType: &sfxMetricTypeCumulativeCounter,
-		}
-		c := int64(inDp.Count())
-		countPt.Value.IntValue = &c
-		out = append(out, &countPt)
-
-		sumPt := sfxpb.DataPoint{
-			Metric:     name,
-			Timestamp:  ts,
-			Dimensions: dims,
-			MetricType: &sfxMetricTypeCumulativeCounter,
-		}
-		sum := inDp.Sum()
-		sumPt.Value.DoubleValue = &sum
-		out = append(out, &sumPt)
-
-		qvs := inDp.QuantileValues()
-		for j := 0; j < qvs.Len(); j++ {
-			qPt := sfxpb.DataPoint{
-				Metric:     name + "_quantile",
-				Timestamp:  ts,
-				MetricType: &sfxMetricTypeGauge,
-			}
-			qv := qvs.At(j)
-			qdim := sfxpb.Dimension{
-				Key:   "quantile",
-				Value: strconv.FormatFloat(qv.Quantile(), 'f', -1, 64),
-			}
-			qPt.Dimensions = append(dims, &qdim)
-			v := qv.Value()
-			qPt.Value.DoubleValue = &v
-			out = append(out, &qPt)
-		}
-	}
-	return out
-}
-
-func convertNumberDatapoints(in pdata.NumberDataPointSlice, basePoint *sfxpb.DataPoint, extraDims []*sfxpb.Dimension) []*sfxpb.DataPoint {
-	out := make([]*sfxpb.DataPoint, 0, in.Len())
-
-	for i := 0; i < in.Len(); i++ {
-		inDp := in.At(i)
-
-		dp := *basePoint
-		dp.Timestamp = timestampToSignalFx(inDp.Timestamp())
-		dp.Dimensions = attributesToDimensions(inDp.Attributes(), extraDims)
-
-		switch inDp.Type() {
-		case pdata.MetricValueTypeInt:
-			val := inDp.IntVal()
-			dp.Value.IntValue = &val
-		case pdata.MetricValueTypeDouble:
-			val := inDp.DoubleVal()
-			dp.Value.DoubleValue = &val
-		}
-
-		out = append(out, &dp)
-	}
-	return out
-}
-
-func makeBaseDataPoint(m pdata.Metric) *sfxpb.DataPoint {
-	return &sfxpb.DataPoint{
-		Metric:     m.Name(),
-		MetricType: fromMetricDataTypeToMetricType(m),
-	}
-}
-
-func fromMetricDataTypeToMetricType(metric pdata.Metric) *sfxpb.MetricType {
-	switch metric.DataType() {
-
-	case pdata.MetricDataTypeGauge:
-		return &sfxMetricTypeGauge
-
-	case pdata.MetricDataTypeSum:
-		if !metric.Sum().IsMonotonic() {
-			return &sfxMetricTypeGauge
-		}
-		if metric.Sum().AggregationTemporality() == pdata.MetricAggregationTemporalityDelta {
-			return &sfxMetricTypeCounter
-		}
-		return &sfxMetricTypeCumulativeCounter
-
-	case pdata.MetricDataTypeHistogram:
-		if metric.Histogram().AggregationTemporality() == pdata.MetricAggregationTemporalityDelta {
-			return &sfxMetricTypeCounter
-		}
-		return &sfxMetricTypeCumulativeCounter
-	}
-
-	return nil
-}
-
-func convertHistogram(histDPs pdata.HistogramDataPointSlice, basePoint *sfxpb.DataPoint, extraDims []*sfxpb.Dimension) []*sfxpb.DataPoint {
-	var out []*sfxpb.DataPoint
-
-	for i := 0; i < histDPs.Len(); i++ {
-		histDP := histDPs.At(i)
-		ts := timestampToSignalFx(histDP.Timestamp())
-
-		countDP := *basePoint
-		countDP.Metric = basePoint.Metric + "_count"
-		countDP.Timestamp = ts
-		countDP.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
-		count := int64(histDP.Count())
-		countDP.Value.IntValue = &count
-
-		sumDP := *basePoint
-		sumDP.Timestamp = ts
-		sumDP.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
-		sum := histDP.Sum()
-		sumDP.Value.DoubleValue = &sum
-
-		out = append(out, &countDP, &sumDP)
-
-		bounds := histDP.ExplicitBounds()
-		counts := histDP.BucketCounts()
-
-		// Spec says counts is optional but if present it must have one more
-		// element than the bounds array.
-		if len(counts) > 0 && len(counts) != len(bounds)+1 {
-			continue
-		}
-
-		for j, c := range counts {
-			bound := infinityBoundSFxDimValue
-			if j < len(bounds) {
-				bound = float64ToDimValue(bounds[j])
-			}
-
-			dp := *basePoint
-			dp.Metric = basePoint.Metric + "_bucket"
-			dp.Timestamp = ts
-			dp.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
-			dp.Dimensions = append(dp.Dimensions, &sfxpb.Dimension{
-				Key:   upperBoundDimensionKey,
-				Value: bound,
-			})
-			cInt := int64(c)
-			dp.Value.IntValue = &cInt
-
-			out = append(out, &dp)
-		}
-	}
-
-	return out
-}
-
 func filterKeyChars(str string, nonAlphanumericDimChars string) string {
 	filterMap := func(r rune) rune {
 		if unicode.IsLetter(r) || unicode.IsDigit(r) || strings.ContainsRune(nonAlphanumericDimChars, r) {
@@ -327,16 +118,6 @@ func filterKeyChars(str string, nonAlphanumericDimChars string) string {
 	}
 
 	return strings.Map(filterMap, str)
-}
-
-func float64ToDimValue(f float64) string {
-	// Parameters below are the same used by Prometheus
-	// see https://github.com/prometheus/common/blob/b5fe7d854c42dc7842e48d1ca58f60feae09d77b/expfmt/text_create.go#L450
-	// SignalFx agent uses a different pattern
-	// https://github.com/signalfx/signalfx-agent/blob/5779a3de0c9861fa07316fd11b3c4ff38c0d78f0/internal/monitors/prometheusexporter/conversion.go#L77
-	// The important issue here is consistency with the exporter, opting for the
-	// more common one used by Prometheus.
-	return strconv.FormatFloat(f, 'g', -1, 64)
 }
 
 // resourceToDimensions will return a set of dimension from the
