+++ exporter/signalfxexporter/internal/translation/from_metrics.go
@@ -1,0 +1,254 @@
+// Copyright OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package translation // import "github.com/open-telemetry/opentelemetry-collector-contrib/exporter/signalfxexporter/internal/translation"
+
+import (
+	"math"
+	"strconv"
+
+	sfxpb "github.com/signalfx/com_signalfx_metrics_protobuf/model"
+	"go.opentelemetry.io/collector/model/pdata"
+)
+
+var (
+	// Some standard dimension keys.
+	// upper bound dimension key for histogram buckets.
+	upperBoundDimensionKey = "upper_bound"
+
+	// infinity bound dimension value is used on all histograms.
+	infinityBoundSFxDimValue = float64ToDimValue(math.Inf(1))
+)
+
+func fromMetric(metric pdata.Metric, extraDimensions []*sfxpb.Dimension) []*sfxpb.DataPoint {
+	// TODO: Figure out some efficient way to know how many datapoints there
+	// will be in the given metric.
+	var dps []*sfxpb.DataPoint
+
+	basePoint := &sfxpb.DataPoint{
+		Metric:     metric.Name(),
+		MetricType: fromMetricTypeToMetricType(metric),
+	}
+
+	switch metric.DataType() {
+	case pdata.MetricDataTypeNone:
+		return nil
+	case pdata.MetricDataTypeGauge:
+		dps = convertNumberDataPoints(metric.Gauge().DataPoints(), basePoint, extraDimensions)
+	case pdata.MetricDataTypeSum:
+		dps = convertNumberDataPoints(metric.Sum().DataPoints(), basePoint, extraDimensions)
+	case pdata.MetricDataTypeHistogram:
+		dps = convertHistogram(metric.Histogram().DataPoints(), basePoint, extraDimensions)
+	case pdata.MetricDataTypeSummary:
+		dps = convertSummaryDataPoints(metric.Summary().DataPoints(), metric.Name(), extraDimensions)
+	}
+
+	return dps
+}
+
+func fromMetricTypeToMetricType(metric pdata.Metric) *sfxpb.MetricType {
+	switch metric.DataType() {
+
+	case pdata.MetricDataTypeGauge:
+		return &sfxMetricTypeGauge
+
+	case pdata.MetricDataTypeSum:
+		if !metric.Sum().IsMonotonic() {
+			return &sfxMetricTypeGauge
+		}
+		if metric.Sum().AggregationTemporality() == pdata.MetricAggregationTemporalityDelta {
+			return &sfxMetricTypeCounter
+		}
+		return &sfxMetricTypeCumulativeCounter
+
+	case pdata.MetricDataTypeHistogram:
+		if metric.Histogram().AggregationTemporality() == pdata.MetricAggregationTemporalityDelta {
+			return &sfxMetricTypeCounter
+		}
+		return &sfxMetricTypeCumulativeCounter
+	}
+
+	return nil
+}
+
+func convertNumberDataPoints(in pdata.NumberDataPointSlice, basePoint *sfxpb.DataPoint, extraDims []*sfxpb.Dimension) []*sfxpb.DataPoint {
+	out := make([]*sfxpb.DataPoint, 0, in.Len())
+
+	for i := 0; i < in.Len(); i++ {
+		inDp := in.At(i)
+
+		dp := *basePoint
+		dp.Timestamp = timestampToSignalFx(inDp.Timestamp())
+		dp.Dimensions = attributesToDimensions(inDp.Attributes(), extraDims)
+
+		switch inDp.Type() {
+		case pdata.MetricValueTypeInt:
+			val := inDp.IntVal()
+			dp.Value.IntValue = &val
+		case pdata.MetricValueTypeDouble:
+			val := inDp.DoubleVal()
+			dp.Value.DoubleValue = &val
+		}
+
+		out = append(out, &dp)
+	}
+	return out
+}
+
+func convertHistogram(histDPs pdata.HistogramDataPointSlice, basePoint *sfxpb.DataPoint, extraDims []*sfxpb.Dimension) []*sfxpb.DataPoint {
+	var out []*sfxpb.DataPoint
+
+	for i := 0; i < histDPs.Len(); i++ {
+		histDP := histDPs.At(i)
+		ts := timestampToSignalFx(histDP.Timestamp())
+
+		countDP := *basePoint
+		countDP.Metric = basePoint.Metric + "_count"
+		countDP.Timestamp = ts
+		countDP.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
+		count := int64(histDP.Count())
+		countDP.Value.IntValue = &count
+
+		sumDP := *basePoint
+		sumDP.Timestamp = ts
+		sumDP.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
+		sum := histDP.Sum()
+		sumDP.Value.DoubleValue = &sum
+
+		out = append(out, &countDP, &sumDP)
+
+		bounds := histDP.ExplicitBounds()
+		counts := histDP.BucketCounts()
+
+		// Spec says counts is optional but if present it must have one more
+		// element than the bounds array.
+		if len(counts) > 0 && len(counts) != len(bounds)+1 {
+			continue
+		}
+
+		for j, c := range counts {
+			bound := infinityBoundSFxDimValue
+			if j < len(bounds) {
+				bound = float64ToDimValue(bounds[j])
+			}
+
+			dp := *basePoint
+			dp.Metric = basePoint.Metric + "_bucket"
+			dp.Timestamp = ts
+			dp.Dimensions = attributesToDimensions(histDP.Attributes(), extraDims)
+			dp.Dimensions = append(dp.Dimensions, &sfxpb.Dimension{
+				Key:   upperBoundDimensionKey,
+				Value: bound,
+			})
+			cInt := int64(c)
+			dp.Value.IntValue = &cInt
+
+			out = append(out, &dp)
+		}
+	}
+
+	return out
+}
+
+func convertSummaryDataPoints(
+	in pdata.SummaryDataPointSlice,
+	name string,
+	extraDims []*sfxpb.Dimension,
+) []*sfxpb.DataPoint {
+	out := make([]*sfxpb.DataPoint, 0, in.Len())
+
+	for i := 0; i < in.Len(); i++ {
+		inDp := in.At(i)
+
+		dims := attributesToDimensions(inDp.Attributes(), extraDims)
+		ts := timestampToSignalFx(inDp.Timestamp())
+
+		countPt := sfxpb.DataPoint{
+			Metric:     name + "_count",
+			Timestamp:  ts,
+			Dimensions: dims,
+			MetricType: &sfxMetricTypeCumulativeCounter,
+		}
+		c := int64(inDp.Count())
+		countPt.Value.IntValue = &c
+		out = append(out, &countPt)
+
+		sumPt := sfxpb.DataPoint{
+			Metric:     name,
+			Timestamp:  ts,
+			Dimensions: dims,
+			MetricType: &sfxMetricTypeCumulativeCounter,
+		}
+		sum := inDp.Sum()
+		sumPt.Value.DoubleValue = &sum
+		out = append(out, &sumPt)
+
+		qvs := inDp.QuantileValues()
+		for j := 0; j < qvs.Len(); j++ {
+			qPt := sfxpb.DataPoint{
+				Metric:     name + "_quantile",
+				Timestamp:  ts,
+				MetricType: &sfxMetricTypeGauge,
+			}
+			qv := qvs.At(j)
+			qdim := sfxpb.Dimension{
+				Key:   "quantile",
+				Value: strconv.FormatFloat(qv.Quantile(), 'f', -1, 64),
+			}
+			qPt.Dimensions = append(dims, &qdim)
+			v := qv.Value()
+			qPt.Value.DoubleValue = &v
+			out = append(out, &qPt)
+		}
+	}
+	return out
+}
+
+func attributesToDimensions(attributes pdata.AttributeMap, extraDims []*sfxpb.Dimension) []*sfxpb.Dimension {
+	dimensions := make([]*sfxpb.Dimension, len(extraDims), attributes.Len()+len(extraDims))
+	copy(dimensions, extraDims)
+	if attributes.Len() == 0 {
+		return dimensions
+	}
+	dimensionsValue := make([]sfxpb.Dimension, attributes.Len())
+	pos := 0
+	attributes.Range(func(k string, v pdata.AttributeValue) bool {
+		dimensionsValue[pos].Key = k
+		dimensionsValue[pos].Value = v.AsString()
+		dimensions = append(dimensions, &dimensionsValue[pos])
+		pos++
+		return true
+	})
+	return dimensions
+}
+
+// Is equivalent to strconv.FormatFloat(f, 'g', -1, 64), but hardcodes a few common cases for increased efficiency.
+func float64ToDimValue(f float64) string {
+	// Parameters below are the same used by Prometheus
+	// see https://github.com/prometheus/common/blob/b5fe7d854c42dc7842e48d1ca58f60feae09d77b/expfmt/text_create.go#L450
+	// SignalFx agent uses a different pattern
+	// https://github.com/signalfx/signalfx-agent/blob/5779a3de0c9861fa07316fd11b3c4ff38c0d78f0/internal/monitors/prometheusexporter/conversion.go#L77
+	// The important issue here is consistency with the exporter, opting for the
+	// more common one used by Prometheus.
+	switch {
+	case f == 0:
+		return "0"
+	case f == 1:
+		return "1"
+	case math.IsInf(f, +1):
+		return "+Inf"
+	default:
+		return strconv.FormatFloat(f, 'g', -1, 64)
+	}
+}
